+++ Src/FluentAssertions/Equivalency/SelfReferenceEquivalencyAssertionOptions.cs
@@ -600,7 +600,8 @@ namespace FluentAssertions.Equivalency
         }
 
         /// <summary>
-        /// Ensures records by default are compared by value instead of their members.
+        /// Ensures records by default are compared by their members even though they override 
+        /// the <see cref="object.Equals(object)" /> method.
         /// </summary>
         /// <remarks>
         /// This is the default.